src/ZenCentral/requirements.txt
django >=1.9, <2
djangorestframework>=3.6.2
django_enumfields>=0.9.0
django-taggit>=0.22.2
django-taggit-serializer>=0.1.5
django_jsonfield>=1.0.1